refactor: component rendering
