+++ src/components/Navbar/NavbarButtons/index.ts
@@ -1,4 +1,3 @@
 export * from './NavbarButtons';
 export * from './NavbarButtons.style';
-export * from './WalletManagementButtons';
-export * from './WalletManagementButtons.style';
+export * from './RedirectToApp';
